Final_Project/Part_2/requirements.txt
pyspark==3.4.0
apache-airflow==2.7.0
apache-airflow-providers-apache-spark==4.1.0
requests==2.31.0